requirements.txt
torch
torchvision
pytorch-lightning
streamlit
transformers
datasets
pandas
scikit-learn
faiss-cpu
altair
einops

